Clarifies that wasix toolchain is installed when you first run a command that requires it
You can not verify wasix toolchain is installed right after running cargo install cargo-wasix because this command does not install wasix toolchain. You need to run a command that requires the toolchain for it to be installed. Just after that you could verify the toolchain is installed with rustup toolchain list